refractor: pdfme functions -> pdfmeUtil
