+++ frontend/src/components/Feature/History/HistoryManageTable.tsx
@@ -6,7 +6,6 @@ import React from 'react';
 import { HistorySnapshotBundledData, HistorySnapshotResponse, MakeActiveSnapshotRequestDto, TrackResponse } from '../../../util/types';
 import { useCustomAlert } from '../../Hooks/useCustomAlert';
 import { mapTemplateInputsToTemplateViewerFromSnapshot, templateInputsFromHistorySnapshot } from '../../../util/dataHelpers';
-import { getPlugins } from '../../../util/helper';
 import { SpinnerDefault } from '../../MenuItems/Loaders/SpinnerDefault';
 import { ConfirmationPopup } from '../../MenuItems/Popups/ConfirmationPopup';
 import { AlertPopup, CustomAlertPopupProps, PopupType } from '../../MenuItems/Popups/AlertPopup';
@@ -22,6 +21,7 @@ import { CloseWindowIcon } from '../../MenuItems/Icons/CloseWindowIcon';
 import { getFontsData } from '../../../util/fontsUtil';
 import { utcFormatter } from '../../../util/datesUtil';
 import { delay } from '../../../util/timeUtil';
+import { getPlugins } from '../../../util/pdfmeUtil';
 
 type Props = {
     getHistory: () => void;
