+++ frontend/src/pages/Verifcation/VerificationPage.tsx
@@ -5,13 +5,13 @@ import { HistorySnapshotResponse } from "../../util/types";
 import { SpinnerDefault } from "../../components/MenuItems/Loaders/SpinnerDefault";
 import { useEffect, useRef, useState } from "react";
 import { Form, Viewer } from "@pdfme/ui";
-import { getPlugins } from "../../util/helper";
 import { mapTemplateInputsToTemplateViewerFromSnapshot, templateInputsFromHistorySnapshot } from "../../util/dataHelpers";
 import { PublishButton } from '../../components/MenuItems/Buttons/PublishButton';
 import { NextIcon } from '../../components/MenuItems/Icons/NextIcon';
 import logoBlack from '/icons/logoBlack.png'
 import { getFontsData } from '../../util/fontsUtil';
 import { generatePDFDownload } from '../../util/pdfGenerationUtil';
+import { getPlugins } from '../../util/pdfmeUtil';
 
 type Props = {
     getHistoryByVerificationCode: (verificationCode: string) => void;
